add CVE-2022-30525 CVE-2022-33891 CVE-2021-24284 2022-07-21

+++ pocs_go/go_poc_check.go
@@ -283,7 +283,7 @@ func init() {
 					close(lib.PocCheck_pipe)
 					return
 				}
-				var f01 float32 = float32(nCnt/nMax) * 100
+				var f01 float32 = float32(nCnt) / float32(nMax) * float32(100)
 				fmt.Printf(" Asynchronous go PoCs detection task %%%0.2f ....\r", f01)
 				<-time.After(time.Duration(1) * time.Second)
 				nCnt += 1
